fix empty XMake configuration selected issue
